requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.0.1
openai>=1.40.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.7.4
pypdf>=4.0.0
